#314
- chips: правки структуры раздела (major)
- chips: Chips.Item помечен как deprecated (patch)
