+++ src/chips/index.ts
@@ -1,0 +1,2 @@
+export { Chips, ChipsProps } from './chips';
+export { ChipsItem, ChipsItemProps } from './chips-item';
